requirements.txt
pyarrow==10.0.1
streamlit==1.22.0
snowflake-sqlalchemy==1.4.7
SQLAlchemy==1.4.48
streamlit==1.22.0
streamlit-chat==0.0.2.2
Python==3.11.0